refactor(ndt_scan_matcher): fixed as pointed out by linter (#5740)
Fixed as pointed out by linter

+++ localization/ndt_scan_matcher/include/ndt_scan_matcher/ndt_scan_matcher_core.hpp
@@ -186,8 +186,8 @@ private:
   double converged_param_transform_probability_;
   double converged_param_nearest_voxel_transformation_likelihood_;
 
-  int initial_estimate_particles_num_;
-  int n_startup_trials_;
+  int64_t initial_estimate_particles_num_;
+  int64_t n_startup_trials_;
   double lidar_topic_timeout_sec_;
   double initial_pose_timeout_sec_;
   double initial_pose_distance_tolerance_m_;
@@ -217,7 +217,7 @@ private:
   bool use_dynamic_map_loading_;
 
   // The execution time which means probably NDT cannot matches scans properly
-  int critical_upper_bound_exe_time_ms_;
+  int64_t critical_upper_bound_exe_time_ms_;
 };
 
 #endif  // NDT_SCAN_MATCHER__NDT_SCAN_MATCHER_CORE_HPP_
